+++ packages/aimd-recorder/src/components/index.ts
@@ -1,14 +1,17 @@
 /**
  * AIMD Editor Components
- * 
+ *
  * Placeholder for component exports
  * Components will be migrated from apps/web/src/components/custom/aimd
  */
 
+import { defineAsyncComponent } from "vue"
 import AimdRecorder from "./AimdRecorder.vue"
 import AimdQuizRecorder from "./AimdQuizRecorder.vue"
 
-export { AimdRecorder, AimdQuizRecorder }
+const AimdDnaSequenceField = defineAsyncComponent(() => import("./AimdDnaSequenceField.vue"))
+
+export { AimdRecorder, AimdQuizRecorder, AimdDnaSequenceField }
 
 /**
  * @deprecated Use `AimdRecorder` instead.